feat: migrate the rest of the shared selectors
